refactor(todo): 移除今日待办功能并新增路径管理支持
- 删除 `internal/tui/models/todo/today.go` 和对应的 CLI 命令 `cmd/todo/todo_today.go`。
- 在 `cmd/group/icons.go` 和 `internal/tui/styles/icons.go` 中定义 Group 和 CLI 专用图标。
- 新增 `internal/models/entity/personal_path.go` 和 `internal/models/personal_path_model.go`，实现个人路径管理，包括路径创建、更新、查询及删除。
- 增加 `internal/database/snowflake.go` 以支持雪花算法生成全局唯一 ID。
- 重构数据模型，统一将 `ID` 改为 `int64` 类型，并优化关联字段命名。
- 在 `components/badge.go` 和 `components/spinner.go` 中更新状态、优先级与加载指示器格式，替换 Emoji 为字符符号。
- 修订 `main.go`，通过子命令自动注册路径和组相关命令。
- 更新 Go 模块依赖，添加 `github.com/bwmarrin/snowflake` 用于全局 ID 生成。

+++ cmd/memory/memory_delete.go
@@ -13,7 +13,6 @@ import (
 var memoryDeleteID int
 
 // memoryDeleteCmd 删除记忆
-// 嘿嘿~ 删除指定的记忆！🗑️
 var memoryDeleteCmd = &cobra.Command{
 	Use:   "delete",
 	Short: "删除记忆",
@@ -30,7 +29,7 @@ var memoryDeleteCmd = &cobra.Command{
 		defer bs.Shutdown()
 
 		handler := handlers.NewMemoryHandler(bs)
-		if err := handler.Delete(bs.Context(), uint(memoryDeleteID)); err != nil {
+		if err := handler.Delete(bs.Context(), int64(memoryDeleteID)); err != nil {
 			cli.PrintError(err.Error())
 			os.Exit(1)
 		}
