+++ internal/models/dto/plan_dto.go
@@ -3,18 +3,16 @@ package dto
 import "time"
 
 // PlanCreateDTO 创建计划请求
-// 嘿嘿~ 用于创建新计划的数据传输对象！💖
 type PlanCreateDTO struct {
 	Title       string `json:"title"`
 	Description string `json:"description"` // 摘要
-	Content     string `json:"content"`     // 详细内容（新增）
+	Content     string `json:"content"`     // 详细内容
 	Scope       string `json:"scope"`       // personal/group/global（默认 group）
 }
 
 // PlanUpdateDTO 更新计划请求
-// 呀~ 用于更新已有计划的数据传输对象！✨
 type PlanUpdateDTO struct {
-	ID          uint       `json:"id"`
+	ID          int64      `json:"id"`
 	Title       *string    `json:"title,omitempty"`
 	Description *string    `json:"description,omitempty"`
 	Content     *string    `json:"content,omitempty"`
@@ -26,14 +24,13 @@ type PlanUpdateDTO struct {
 
 // PlanProgressDTO 更新计划进度请求
 type PlanProgressDTO struct {
-	ID       uint `json:"id"`
-	Progress int  `json:"progress"` // 0-100
+	ID       int64 `json:"id"`
+	Progress int   `json:"progress"` // 0-100
 }
 
 // PlanResponseDTO 计划响应
-// 嘿嘿~ 用于返回计划详情的数据传输对象！💖
 type PlanResponseDTO struct {
-	ID          uint         `json:"id"`
+	ID          int64        `json:"id"`
 	Title       string       `json:"title"`
 	Description string       `json:"description"` // 摘要
 	Content     string       `json:"content"`     // 详细内容
@@ -49,9 +46,8 @@ type PlanResponseDTO struct {
 }
 
 // PlanListDTO 计划列表项
-// 呀~ 用于列表展示的简化计划数据！✨
 type PlanListDTO struct {
-	ID          uint   `json:"id"`
+	ID          int64  `json:"id"`
 	Title       string `json:"title"`
 	Description string `json:"description"` // 摘要用于列表展示
 	Status      string `json:"status"`
@@ -61,7 +57,7 @@ type PlanListDTO struct {
 
 // SubTaskDTO 子任务 DTO
 type SubTaskDTO struct {
-	ID          uint      `json:"id"`
+	ID          int64     `json:"id"`
 	Title       string    `json:"title"`
 	Description string    `json:"description"`
 	Status      string    `json:"status"`
@@ -73,14 +69,14 @@ type SubTaskDTO struct {
 
 // SubTaskCreateDTO 创建子任务请求
 type SubTaskCreateDTO struct {
-	PlanID      uint   `json:"plan_id"`
+	PlanID      int64  `json:"plan_id"`
 	Title       string `json:"title"`
 	Description string `json:"description"`
 }
 
 // SubTaskUpdateDTO 更新子任务请求
 type SubTaskUpdateDTO struct {
-	ID          uint    `json:"id"`
+	ID          int64   `json:"id"`
 	Title       *string `json:"title,omitempty"`
 	Description *string `json:"description,omitempty"`
 	Status      *string `json:"status,omitempty"`
